requirements.txt
flask
tensorflow
pillow
numpy
python-dotenv
gunicorn
pandas
scikit-learn
